removebg/requirements.txt
rembg==2.0.50
Pillow==9.5.0
types-Pillow==9.5.0.5
